Fix some warning

+++ src/public/FaDateTime.cpp
@@ -236,7 +236,7 @@ int operator -(const FaDateTime& left, const FaDateTime& right)
 bool FaDateTime::verifyFormat()
 {
     //检查时间字符串是否全都是数字
-    for (int i=0; i<m_strDateTime.size(); i++)
+    for (size_t i=0; i<m_strDateTime.size(); i++)
     {
         if (!isdigit(m_strDateTime[i]))
         {
